new file:   main.c 	new file:   prg 	modified:   read_hex.c 	modified:   read_hex.h

+++ read_hex.h
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #define WORDS       32768
 
 static unsigned char data[WORDS];
@@ -6,6 +8,6 @@ void readHexProgram();
 void printHexContents(uint8_t address_locations);
 void printHexData(uint8_t size);
 void writeData(uint8_t size);
-void writeAddress(uint8_t size);
+void writeAddress(uint32_t size);
 int setupWiringPi();
 void setupPins();
